feat: move the `actions` folder to be a subfolder of `lib`
