+++ components/courses/edit-course.tsx
@@ -27,7 +27,7 @@ import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import { toast } from "sonner";
 import { Pencil } from "lucide-react";
-import { updateCourse } from "@/actions/courses";
+import { updateCourse } from "@/lib/actions/courses";
 import type { Course } from "@/generated/prisma/client";
 
 const courseFormSchema = z.object({
