+++ components/dashboard/project-dialog.tsx
@@ -14,7 +14,7 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
 import { toast } from "sonner";
-import { createProject } from "@/actions/projects";
+import { createProject } from "@/lib/actions/projects";
 
 interface ProjectDialogProps {
   open: boolean;
