+++ components/profile/profile-form.tsx
@@ -2,7 +2,7 @@
 
 import { useTransition } from "react";
 import { toast } from "sonner";
-import { updateProfile } from "@/actions/user";
+import { updateProfile } from "@/lib/actions/user";
 import { Button } from "@/components/ui/button";
 import { CardContent, CardFooter } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
